requirements.txt
aiosqlite==0.21.0
alembic==1.16.4
anyio==4.10.0
asyncpg==0.29.0
certifi==2025.8.3
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-telegram-bot==21.6
psycopg2-binary==2.9.9 ; platform_system != "Windows"
sniffio==1.3.1
SQLAlchemy==2.0.32
typing_extensions==4.14.1
aiohttp
